21.1.Embedded Relation {Embedded CheeseListing into User}
Embedded CheeseListing into User

When two resources are related to each other, this can be expressed in two different ways in an API. The first is with IRIs - basically a "link" to the other resource. We can't see the data for the related CheeseListing, but if we need it, we could make a second request to this URL and... boom! We've got it.

But, for performance purposes, you might say:

You know what? I don't want to have to make one request to fetch user data and then one more request to get the data for each cheese listing they own: I want to get it all at once!

And that describes the second way of expressing a relationship: instead of just returning a link to a cheese listing, you can embed its data right inside!

As a reminder, when we normalize a User, we include everything in the user:read group. So that means $email, $username and $cheeseListings, which is why that property shows up at all.

To make this property return data, instead of just an IRI, here's what you need to do: go into the related entity - so CheeseListing - and add this user:read group to at least one property. For example, add user:read above $title... and how about also above $price.

Let's see what happens! We don't even need to refresh, just Execute. Woh! Instead of an array of strings, it's now an array of objects! Well, this user only owns one CheeseListing, but you get the idea. Each item has the standard @type and @id plus whatever properties we added to the group: title and price.

It's beautifully simple: the serializer knows to serialize all fields in the user:read group. It first looks at User and finds email, username and cheeseListings. It then keeps going and, inside of CheeseListing, finds that group on title and price.

Relation Strings vs Objects

This means that each relation property may be a string - the IRI - or an object. And an API client can tell the difference. If you get back an object, you know it will have @id, @type and some other data properties. If you get back a string, you know it's an IRI that you can use to go get the real data.