21.2.Embedded Relation {Embedding User into CheeseListing}
Embedding User into CheeseListing

We can do the same thing on the other side of the relationship. Use the docs to get the CheeseListing with id = 1. Yep! The owner property is a string. But it might be convenient for the CheeseListing JSON to at least contain the username of the owner... so we don't need to go fetch the entire User just to display who owns it.

Inside CheeseListing, the normalization process will serializer everything in the cheese_listing:read group. Copy that. The owner property, of course, already has this group above it, which is why we see it in our API. Inside User, find $username... and add cheese_listing:read to that.

Let's try this thing! Move back over and... Execute! And... ha! Perfect! It expands to an object and includes the username.

Does it work if we GET the collection of cheese listings? Try it out! Well... ok, there's only one CheeseListing in the database right now, but of course! It embeds the owner in the same way.