21.3.Embedded Relation {Embedding Data Only on when GETing a Single Item}
Embedding Data Only on when GETing a Single Item

So... about that... new challenge! What if we want to embed the owner data when I fetch a single CheeseListing... but, to keep the response from being gigantic... we don't want to embed the data when we fetch the collection. Is that possible?

Totally! Again, for CheeseListing, when we normalize, we include everything in the cheese_listing:read group. That's true regardless of whether we're GETting the collection of cheese listings or just GETting a single item. But, tons of things - including groups - can be changed on an operation-by-operation basis.

For example, under itemOperations, break the get operation configuration onto multiple lines and add normalization_context. One of the tricky things with the config here is that the top-level keys are lower camel case, like normalizationContext. But deeper keys are usually snake case, like normalization_context. That... can be a little inconsistent - and it's easy to mess these up. Be careful.

Anyways, the goal is to override the normalization context, but only for this one operation. Set this to the normal groups and another array. Inside, we're going to say:

Hey! When you are getting a single item, I want to include all of the properties that have the cheese_listing:read group like normal. But I also want to include any properties in a new cheese_listing:item:get group.

We'll talk more about it later - but I'm using a specific naming convention for this operation-specific group - the "entity name", colon, item or collection, colon, then the HTTP method - get, post, put, etc.

If we re-fetch a single CheeseListing.... it makes no difference: we're including a new group for serialization - yaaaay - but nothing is in the new group.

Here's the magic. Copy the new group name, open User, and above the $username property, replace cheese_listing:read with cheese_listing:item:get.

That's it! Move back to the documentation and fetch a single CheeseListing. And... perfect - it still embeds the owner - there's the username. But now, close that up and go to the GET collection endpoint. Execute! Yes! Owner is back to an IRI!

These serialization groups can get a little complex to think about, but wow are they powerful.

Next... when we fetch a CheeseListing, some of the owner's data is embedded into the response. So... I have kind of a crazy question: when we're updating a CheeseListing... could we also update some data on the owner by sending embedded data? Um... yea! That's next.